Added some very basic Logic to break up loops
Added a very crude way to break up Cycles in the given Graph. However this is a very basic principle
and the Rest of the Codebase then displays them without any regard for the Fact that it was previously a Cycle/Back-Edge.

This was only mainly added to avoid crashes when encountering Cycles and display something but it is not
easily understandable or understandable at all