The cycles are now being rendered properly
Graphs that contain cycles can now be displayed properly although they are not always
layed out optimally, at least the work.
However the code is still really messy and needs a bit of a rework/cleanup to make it more
maintanable and easier to fix the current problems that are still around.